+++ components/analisis/templates/layout_analisis.html
@@ -7,20 +7,38 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<link href="https://fonts.googleapis.com/css?family=Barlow+Condensed|Francois+One|Pacifico" rel="stylesheet">
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+	<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
 	<link rel="stylesheet" type="text/css" href="{% static 'style.css' %}">
 	<!-- <link rel="shortcut icon" href="../browser.ico"/> -->
 </head>
 
 <body>
+	<div>
+		<nav class="navbar navbar-default navigation">
+			<div class="container-fluid micro-top-space">
+				<div class="navbar-header">
+					<a class="navbar-brand pacifico" href="">Automatic Cell Detector</a>
+				</div>
+				<ul class="nav navbar-nav">
+					<li><a href="{% url 'index' %}">Home</a></li>
+					<li class="active"><a href="{% url 'analisis' %}">Análisis</a></li>
+					<li><a href="{% url 'Experimento_list' %}">Experimentos</a></li>
+					<li><a href="{% url 'tratamientos_index' %}">Tratamientos</a></li>
+					<li><a href="{% url 'busqueda_home' %}">Búsquedas</a></li>
+				</ul>
+				<ul class="nav navbar-nav navbar-right">
+					<li><a href="#"><span class="glyphicon glyphicon-user"></span> Hola Usuario</a></li>
+					<li><a href="http://127.0.0.1:8000/"><span class="glyphicon glyphicon-log-in"></span> Salir</a></li>
+				</ul>
+			</div>
+		</nav>
+	</div>
+
 	<div class"container">
 		{% block content %}
 		{% endblock %}
 	</div>
-	<br>
-	<br>
-	<br>
-	<br>
-	<br>
+
 	<footer class="footer top-space blue-bg white" style="position: absolute;margin-bottom: opx;">
 		<div class="container">
 			<div class="row">
@@ -38,11 +56,11 @@
 				<div class="col-sm-6 mini-top-space">
 					<h2 class="pacifico white">Collaborators</h2>
 					<div class="col-sm-6">
-							<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'parma80.png' %}" alt="Smiley face"></a>
-						</div>
-	
-						<div class="col-sm-6">
-							<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'lab.png' %}" alt="Smiley face"></a>
+						<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'parma80.png' %}" alt="Smiley face"></a>
+					</div>
+
+					<div class="col-sm-6">
+						<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'lab.png' %}" alt="Smiley face"></a>
 					</div>
 				</div>
 			</div>
