+++ components/analisis/templates/subir.html
@@ -1,0 +1,27 @@
+{% extends "layout_analisis.html" %}
+
+{% block content %}
+
+<div class="container top-space">
+  <center>
+    <h1>Análisis de imágenes</h1>
+    <form enctype="multipart/form-data" method="post" action="{% url 'analisis' %}">
+      {% csrf_token %}
+      <table>
+        {{ form.as_p}}
+      </table><input type="submit">
+      <img class="img-fluid d-block" id="imagen" src="{{ imgsrc }}">
+    </form>
+
+    <br>
+    <br>
+    <br>
+
+    <div>
+      <li><a class="btn btn-default" href="{% url 'resultados' %}">Analizar</a></li>
+    </div>
+  </center>
+</div>
+
+
+{% endblock %}
