+++ components/analisis/views.py
@@ -13,7 +13,27 @@ import os
 
 from components.experimento.models import Experimento
 
+from components.home.forms import UploadFileForm
+import cloudinary
+
 guardado = False
+
+@login_required
+def analisis(request):
+    imgsrc = "https://uploads-ssl.webflow.com/57e5747bd0ac813956df4e96/5aebae14c6d254621d81f826_placeholder.png"
+    form = UploadFileForm(request.POST, request.FILES)
+    
+    if request.method == 'POST':
+        if form.is_valid():
+            #do things
+            imagen = form.cleaned_data['file']
+            cld = cloudinary.uploader.upload(imagen)
+            url = cld['url']
+            request.session['url'] = url
+            return render(request, 'subir.html', {'form': form, 'imgsrc' : url})
+
+
+    return render(request, 'subir.html', {'form': form, 'imgsrc' : imgsrc})
 
 # Create your views here.
 @login_required
