--- components/experimento/templates/experimento_del.html
@@ -1,5 +1,0 @@
-<h1>Experimento Delete</h1>
-<form method="post">{% csrf_token %}
-    Are you sure you want to delete "{{ object }}" ?
-    <input type="submit" value="Submit" />
-</form>
