--- components/home/templates/resultados.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Automatic Cell Detector</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<link href="https://fonts.googleapis.com/css?family=Barlow+Condensed|Francois+One|Pacifico" rel="stylesheet">
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-	<link rel="stylesheet" type="text/css" href="./style.css">
-	<link rel="stylesheet" type="text/css" href="./style_login.css">
-	<!-- <link rel="shortcut icon" href="../browser.ico"/> -->
-</head>
-<body>
-
-	<div>
-	<nav class="navbar navbar-default navigation">
-		<div class="container-fluid micro-top-space">
-    		<div class="navbar-header">
-      			<a class="navbar-brand pacifico" href="">Automatic Cell Detector</a>
-    		</div>
-		    <ul class="nav navbar-nav">
-		      <li ><a href="http://127.0.0.1:8000/index/">Home</a></li>
-		      <li ><a href="http://127.0.0.1:8000/analisis/">Análisis</a></li>
-		      <li class="active"><a href="">Resultados</a></li>
-		    </ul>
-		    <ul class="nav navbar-nav navbar-right">
-		      <li><a href="#"><span class="glyphicon glyphicon-user"></span> Hola Usuario</a></li>
-		      <li><a href="http://127.0.0.1:8000/"><span class="glyphicon glyphicon-log-in"></span> Salir</a></li>
-		    </ul>
-  		</div>
-	</nav>
-	</div>
-
-	<div class="container top-space">
-		<center>		
-			<h1>Resultados</h1>
-			<div class="row">
-				<div class="col-sm-6">
-					<h3>Original</h3> <br>
-					<img src="http://res.cloudinary.com/ivanfelipecp/image/upload/v1536739971/qbb4yyvjzdtstfefnwzv.png" class="img-rounded" width="256" height="256"> 
-				</div>
-				<div class="col-sm-6">
-					<h3>Segmentada</h3> <br>
-					<img src="http://res.cloudinary.com/ivanfelipecp/image/upload/v1536739978/lx4ne37zj1fn41cizep9.png" class="img-rounded" width="256" height="256"> 
-				</div>				
-			</div>
-
-			<form id="formulario" class="top-space">
-				<div class="form-group">
-					<label for="exampleFormControlFile1">Comentario</label> <br>
-					<textarea name="comment" form="formulario" rows="4" cols="50" placeholder="Comente la observación"></textarea>
-				</div>
-
-			<div class="form-group">
-				<label for="sel1">Seleccione el paciente</label>
-				<select class="form-control" id="sel1" style="width: 20%">
-					<option>Stuart Little</option>
-					<option>Iván Calvo</option>
-					<option>Sisokita</option>
-					<option>Adriansito bb</option>
-				</select>
-			</div>
-
-			  <div class="form-group">
-			    <label for="exampleFormControlFile1">Seleccione la fecha</label> <br>
-			    <input type="date" id="">
-			  </div>
-
-			  <div class="form-group">
-			  	<input type="submit" class="btn btn-default">
-			  </div>			  
-			</form>
-		</center>			
-	</div>
-
-	<footer class="footer top-space blue-bg white">
-    	<div class="container">
-        	<div class="row">
-          		<div class="col-sm-6 mini-top-space">
-            		<h1 class="pacifico white">Developed by ISLC S.A &copy;</h1>
-          		</div>
-
-		        <div class="col-sm-3">
-		        	<h2 class="pacifico white">Contact US</h2>
-		        	<a href="#" class="white a-hover">https://www.sleepingdogs.com</a></br>
-		            <h5 class="white">sleepingdogsdesign@gmail.com</h5>
-		            <h5 class="white">+506 888-8888</h5>
-		        </div>
-
-        		<div class="col-sm-3">
-		            <h2 class="pacifico white">About US</h2>
-		            <a href="#" class="white a-hover">Facebook</a></br>
-		            <a href="#" class="white a-hover">Developers</a></br>
-		            <a href="#" class="white a-hover">Linkedin</a></br>
-        		</div>
-       		</div>
-      	</div>
-	</footer>
-
-	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-</body>
-</html>
