--- components/home/templates/starter-template.css
@@ -1,7 +1,0 @@
-body {
-    padding-top: 5rem;
-  }
-  .starter-template {
-    padding: 3rem 1.5rem;
-    text-align: center;
-  }
