+++ components/login/templates/layout_login.html
@@ -1,0 +1,59 @@
+{% load static %}
+
+<html>
+
+<head>
+	<title>Automatic Cell Detector</title>
+	<meta name="viewport" content="width=device-width, initial-scale=1">
+	<link href="https://fonts.googleapis.com/css?family=Barlow+Condensed|Francois+One|Pacifico" rel="stylesheet">
+	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+	<link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
+	<link rel="stylesheet" type="text/css" href="{% static 'style.css' %}">
+	<!-- <link rel="shortcut icon" href="../browser.ico"/> -->
+</head>
+
+<body>
+	<div class"container">
+		{% block content %}
+		{% endblock %}
+	</div>
+	<br>
+	<br>
+	<br>
+	<br>
+	<br>
+	<footer class="footer top-space blue-bg white" style="position: absolute;margin-bottom: opx;">
+		<div class="container">
+			<div class="row">
+				<div class="col-sm-3 mini-top-space">
+					<h1 class="pacifico white">Developed by ISLC S.A &copy;</h1>
+				</div>
+
+				<div class="col-sm-3 mini-top-space">
+					<h2 class="pacifico white">Contact US</h2>
+					<a href="#" class="white a-hover">https://www.islcsa.com</a></br>
+					<h5 class="white">islc@gmail.com</h5>
+					<h5 class="white">+506 888-8888</h5>
+				</div>
+
+				<div class="col-sm-6 mini-top-space">
+					<h2 class="pacifico white">Collaborators</h2>
+					<div class="col-sm-6">
+							<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'parma80.png' %}" alt="Smiley face"></a>
+						</div>
+	
+						<div class="col-sm-6">
+							<a href="https://www.facebook.com/ParmaResearchGroup/"><img src="{% static 'lab.png' %}" alt="Smiley face"></a>
+					</div>
+				</div>
+			</div>
+
+		</div>
+		</div>
+	</footer>
+
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+</body>
+
+</html>
