+++ components/login/templates/login.html
@@ -1,4 +1,4 @@
-{% extends "layout.html" %}
+{% extends "layout_login.html" %}
 
 {% block content %}
 
@@ -32,7 +32,7 @@
       </table>
 
       <input type="submit" value="login">
-      <input type="hidden" name="next" value="/index/">
+      <input type="hidden" name="next" value="/home/">
     </form>
   </center>
 </div>
