+++ components/login/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from django.contrib.auth.views import LoginView, LogoutView
+
+from . import views
+
+urlpatterns = [
+    path('', LoginView.as_view(template_name='login.html'), name='login')
+]
