+++ components/login/views.py
@@ -5,7 +5,3 @@ from django.http import HttpResponse
 from django.shortcuts import render
 from django.conf import settings
 from django.core.files.storage import FileSystemStorage
-
-
-def index(request):
-    return HttpResponse("asdf")
