+++ components/static/style.css
@@ -283,7 +283,7 @@ a{
 
 
 .footer {
-  position: static;
+  position: absolute;
   right: 0;
   bottom: 0;
   left: 0;
